clearing the fields as soons as task is added
